+++ lib/di/category.dart
@@ -3,9 +3,9 @@ import 'dart:io';
 
 import 'package:genshin_mod_manager/data/helper/fsops.dart';
 import 'package:genshin_mod_manager/data/helper/path_op_string.dart';
+import 'package:genshin_mod_manager/di/app_state.dart';
 import 'package:genshin_mod_manager/domain/entity/mod.dart';
 import 'package:genshin_mod_manager/domain/entity/mod_category.dart';
-import 'package:genshin_mod_manager/flow/app_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'category.g.dart';
