+++ lib/di/nahida_store.dart
