+++ lib/ui/route/category/category.dart
@@ -10,13 +10,13 @@ import 'package:genshin_mod_manager/data/helper/mod_switcher.dart';
 import 'package:genshin_mod_manager/data/helper/path_op_string.dart';
 import 'package:genshin_mod_manager/data/repo/akasha.dart';
 import 'package:genshin_mod_manager/data/repo/mod_writer.dart';
+import 'package:genshin_mod_manager/di/app_state.dart';
+import 'package:genshin_mod_manager/di/category.dart';
+import 'package:genshin_mod_manager/di/ini_widget.dart';
+import 'package:genshin_mod_manager/di/mod_card.dart';
 import 'package:genshin_mod_manager/domain/entity/ini.dart';
 import 'package:genshin_mod_manager/domain/entity/mod.dart';
 import 'package:genshin_mod_manager/domain/entity/mod_category.dart';
-import 'package:genshin_mod_manager/flow/app_state.dart';
-import 'package:genshin_mod_manager/flow/category.dart';
-import 'package:genshin_mod_manager/flow/ini_widget.dart';
-import 'package:genshin_mod_manager/flow/mod_card.dart';
 import 'package:genshin_mod_manager/ui/constant.dart';
 import 'package:genshin_mod_manager/ui/util/display_infobar.dart';
 import 'package:genshin_mod_manager/ui/widget/category_drop_target.dart';
