+++ lib/ui/route/welcome.dart
@@ -1,7 +1,7 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/gestures.dart';
+import 'package:genshin_mod_manager/di/app_state.dart';
 import 'package:genshin_mod_manager/domain/constant.dart';
-import 'package:genshin_mod_manager/flow/app_state.dart';
 import 'package:genshin_mod_manager/ui/util/open_url.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
