+++ lib/ui/widget/preset_control.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:genshin_mod_manager/di/preset.dart';
 import 'package:genshin_mod_manager/domain/entity/mod_category.dart';
-import 'package:genshin_mod_manager/flow/preset.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 /// A widget that provides a control for presets.
